Delete LinkedList.c
This was a linked list implementation that I wrote awhile back to use in conjunction with the hash table code that I wrote. However, I've since rewritten the collision resolution portion of the insert code so that it doesn't use a linked list, instead it just does linear probing to find an unused slot in the table.